#80 - refactor to have common REST types
